requirements.txt
argilla==2.1.0
pydantic==2.10.6
langchain==0.2.16
langchain-anthropic==0.1.23
langchain-community==0.2.16
langchain-openai==0.1.7
duckduckgo_search==5.2.0
google-search-results==2.4.2
requests>=2.31
pre-commit==3.8.0
pypdf==4.3.1
PyMuPDF==1.24.9
pytest==8.3.2
python-dotenv==1.0.1
ruff==0.6.4
watchdog==3.0.0
fastapi>=0.115.2,<1.0
uvicorn==0.30.6
python-multipart
cryptography>=3.4
chromadb==0.5.5
langchain-chroma==0.1.2
posthog==2.4.2
